move shop_items into models/ folder
