+++ server/aumo/models/model.go
@@ -1,4 +1,4 @@
-package aumo
+package models
 
 import "time"
